Gates and Policies

+++ app/Http/Controllers/IdeaController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Http\Requests\IdeaRequest;
 use App\Models\Idea;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
     public function index()
     {
+        Gate::authorize('viewAny', Idea::class);
         return view('ideas.index', [
             'ideas' => Auth::user()->ideas()->latest()->get()
         ]);
@@ -17,11 +19,14 @@ class IdeaController extends Controller
 
     public function create()
     {
+        Gate::authorize('create', Idea::class);
         return view('ideas.create');
     }
 
     public function show(Idea $idea)
     {
+        Gate::authorize('view', $idea);
+    
         return view('ideas.show', [
             'idea' => $idea
         ]);
@@ -29,6 +34,7 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+         Gate::authorize('update', $idea);
         return view('ideas.edit', [
             'idea' => $idea
         ]);
@@ -36,6 +42,7 @@ class IdeaController extends Controller
 
     public function store(IdeaRequest $request)
     {
+        Gate::authorize('create', Idea::class);
         $validated = $request->validated();
 
         Auth::user()->ideas()->create([
@@ -48,6 +55,7 @@ class IdeaController extends Controller
 
     public function update(IdeaRequest $request, Idea $idea)
     {
+        Gate::authorize('update', $idea);
         $validated = $request->validated();
 
         $idea->update([
@@ -61,6 +69,7 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        Gate::authorize('delete', $idea);
         $idea->delete();
         return redirect('/ideas');
     }
